# backend/src/main/resources/static/openapi/aidrun-mvp.yaml
openapi: 3.0.3
info:
  title: AidRun MVP v0.3 API
  version: 0.3.0
  description: |
    助盲跑 / AidRun Swift iOS + Spring Boot MVP REST API contract.
    This contract is frozen for the 3-day demo MVP and intentionally excludes
    WebSocket, real SMS, real identity verification, real admin backend,
    real-time tracks, in-app chat, AI assistant, route navigation, payment,
    inventory, Android, and instant-call dispatch.
servers:
  - url: http://localhost:8080
    description: Local Spring Boot backend
  - url: http://192.168.0.100:8080
    description: LAN demo backend example
...
security:
  - bearerAuth: []
paths:
...
  /api/users/me:
    get:
      tags: [Users]
      summary: 获取当前用户
      responses:
        '200':
          description: 当前用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/users/me/active-role:
    patch:
      tags: [Users]
      summary: 切换 activeRole
      description: 共用同一个 JWT；存在 accepted / arrived / in_progress / emergency 订单时禁止切换。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwitchRoleRequest'
      responses:
        '200':
          description: 切换成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '409':
          $ref: '#/components/responses/ActiveOrderRoleSwitchBlocked'
  /api/profiles/blind-runner:
    put:
      tags: [Profiles]
      summary: 盲人资料创建/更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlindRunnerProfileUpsertRequest'
      responses:
        '200':
          description: 已保存盲人资料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlindRunnerProfileDto'
  /api/profiles/volunteer:
    put:
      tags: [Profiles]
      summary: 志愿者资料创建/更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VolunteerProfileUpsertRequest'
      responses:
        '200':
          description: 已保存志愿者资料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolunteerProfileDto'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: 未登录或 token 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: 请先登录
...
    ActiveOrderRoleSwitchBlocked:
      description: 存在活跃订单，禁止切换角色
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: ACTIVE_ORDER_ROLE_SWITCH_BLOCKED
            message: 当前存在进行中的订单，暂时不能切换身份
  schemas:
    ErrorCode:
      type: string
      enum:
        - INVALID_VERIFICATION_CODE
        - PROFILE_INCOMPLETE
        - LOCATION_PERMISSION_REQUIRED
        - ORDER_NOT_FOUND
        - ORDER_ALREADY_ACCEPTED
        - INVALID_ORDER_STATUS
        - ACTIVE_ORDER_ROLE_SWITCH_BLOCKED
        - VOLUNTEER_NOT_AVAILABLE
        - VOLUNTEER_NOT_APPROVED
        - APPOINTMENT_TOO_SOON
        - UNAUTHORIZED
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
    UserRole:
      type: string
      enum: [blind_runner, volunteer]
    VerificationStatus:
      type: string
      enum: [not_submitted, pending, approved, rejected]
    AdminReviewStatus:
      type: string
      enum: [not_submitted, pending, approved, rejected]
...
    UserDto:
      type: object
      required: [id, phoneNumber, roles, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        phoneNumber:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRole'
        activeRole:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserMeResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/UserDto'
        blindRunnerProfile:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/BlindRunnerProfileDto'
        volunteerProfile:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/VolunteerProfileDto'
    SwitchRoleRequest:
      type: object
      required: [activeRole]
      properties:
        activeRole:
          $ref: '#/components/schemas/UserRole'
    EmergencyContactDto:
      type: object
      required: [name, phoneNumber]
      properties:
        name:
          type: string
        phoneNumber:
          type: string
    BlindRunnerProfileUpsertRequest:
      type: object
      required: [nickname, emergencyContact]
      properties:
        nickname:
          type: string
        runningExperience:
          type: string
        emergencyContact:
          $ref: '#/components/schemas/EmergencyContactDto'
    BlindRunnerProfileDto:
      type: object
      required: [id, userId, nickname, emergencyContact, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        nickname:
          type: string
        runningExperience:
          type: string
          nullable: true
        emergencyContact:
          $ref: '#/components/schemas/EmergencyContactDto'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VolunteerProfileUpsertRequest:
      type: object
      required: [nickname]
      properties:
        nickname:
          type: string
    VolunteerProfileDto:
      type: object
      required:
        - id
        - userId
        - nickname
        - phoneNumber
        - verificationStatus
        - adminReviewStatus
        - isAvailable
        - pointsBalance
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        nickname:
          type: string
        phoneNumber:
          type: string
        verificationStatus:
          $ref: '#/components/schemas/VerificationStatus'
        adminReviewStatus:
          $ref: '#/components/schemas/AdminReviewStatus'
        isAvailable:
          type: boolean
        pointsBalance:
          type: integer
          example: 100
...